% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpdist.R
\name{bpdist}
\alias{bpdist}
\title{Distribution of breakpoints}
\usage{
bpdist(topres, ndigits=0, makeplot=FALSE, savePDF=FALSE, barcolor=NULL, outfile, xlabname=NULL)
}
\arguments{
\item{topres}{results from topsegreg() function}

\item{ndigits}{how many digits to be used when rounding. Default is 0 (return integers).}

\item{makeplot}{whether to plot the breakpoint distribution. Default is true.}

\item{savePDF}{whether to save the plot. Default is FALSE.}

\item{outfile}{if savePDF is TRUE then this is the name of the file and path of the plot. Default is breakpoint_distribution.pdf saved in the working directory.}

\item{xlabname}{name of x-axis. Default is Time.}

\item{ndigits}{how many digits to be used when rounding. Default is 0 (return integers).}
}
\value{
The function takes significant genes called from 
the topsegreg() function. For any time point, this function calculates how many genes
have breakpoint at this time point. The output is the numbers of genes sorted
by time point, and optionally a barplot of the breakpoint distribution.
}
\author{
Rhonda Bacher & Ning Leng
}

