% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractpattern.R
\name{extractpattern}
\alias{extractpattern}
\title{extract pattern from segmented regression}
\usage{
extractpattern(segdata, radjcut = 10, pattern = 0)
}
\arguments{
\item{segdata}{segreg object}

\item{radjcut}{only consider genes with adjusted R^2 > radjcut. Default = .5.}

\item{pattern}{vector containing pattern to search genes. If length is one then consider genes with constant pattern.}
}
\value{
genes: name of genes containing pattern
}
\examples{
extractpattern(segdata, pattern = c("up")) #increasing only genes
extractpattern(segdata, pattern = c("up", "down")) #genes with a peak
}
\author{
Rhonda Bacher
}

