% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmarker.R
\name{plotmarker}
\alias{plotmarker}
\title{Plot markers of interest}
\usage{
plotmarker(indata, t.vect = NULL, outfile = NULL, listfeatures, savePDF = FALSE, plotFit = TRUE, fittedreg=NULL,
cond.col=NULL, cond.col.sample=NULL, cond.prefix="day",
yname="normalized expression",par.param = c(3,2),
pdfheight=15, pdfwidth=10)
}
\arguments{
\item{indata}{matrix of normalized expression measures. Rows are genes and columns are samples. The data matrix is expected to be normalized.}

\item{t.vect}{a numerical vector indicates time points. If it is
NULL (default), the time will be assumed to be 1:N in which N is number of equally spaced samples.}

\item{outfile}{output file name; will be ignored if savePDF=F. Default is feature_expressionScatter.}

\item{listfeatures}{a list of genes of interest}

\item{savePDF}{whether output figs to a pdf file.}

\item{plotFit}{whether to plot the segmented regression fitting.}

\item{fittedreg}{segmented regression fitting result; if fittedres is NULL and seg=T, segmented regression will be fitted for each of the genes and it may take longer runtime}

\item{cond.col}{color for each condition, names of this vector should be condition names; if it is null, no legends will be generated}

\item{cond.col.sample}{each sample's color. The vector's length should match number of samples. If it is NULL, expressions will be shown in black.}

\item{cond.prefix}{prefix to be appended to condition name}

\item{yname}{y axis name}

\item{par.param}{mfrow, default c(3,2)}

\item{pdfheight, pdfwidth}{height and width for the pdf file}
}
\value{
plot of gene expression and fitted lines.
}
\examples{
d1 <- rbind(c(rep(1,50),1:50), c(100:1))
rownames(d1) <- c("g1","g2")
plotmarker(d1, listname=c("g1","g2"), pdf = FALSE)
}
\author{
Ning Leng & Rhonda Bacher
}

