\name{SegReg-package}
\alias{SegReg-package}
\alias{SegReg}
\docType{package}
\title{
\packageTitle{SegReg}
}
\description{
\packageDescription{SegReg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SegReg}
\packageIndices{SegReg}
}
\author{
\packageAuthor{SegReg}

Maintainer: \packageMaintainer{SegReg}
}
\references{
}
	\keyword{ package }
\seealso{
}
\examples{
}
