% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg.R
\name{segreg}
\alias{segreg}
\title{segmented regression on a set of genes}
\usage{
segreg(data, meancut=10,maxk=3, t.vect=NULL, min.num.in.seg=5, pvalcut=.1,
               cutdiff=.1, num.try=100,keepfit=FALSE, forceRadj = FALSE, NCores=NULL, makeShinyObject=TRUE)
}
\arguments{
\item{indata}{matrix of normalized expression measures. Rows are genes and columns are samples. The data matrix is expected to be normalized.}

\item{meancut}{default is 10. Genes whose mean is less than meancut will not be considered.}

\item{maxk}{max number of breakpoints to consider.}

\item{t.vect}{a numerical vector indicates time points. If it is
NULL (default), the time will be assumed to be 1:N in which N is number of equally spaced samples.}

\item{min.num.in.seg}{min number of samples within a segment.}

\item{pvalcut}{p-value cutoff. If the p-value of a segment is greater than pvalcut,
this segment will be called as 'no change'.}

\item{cutdiff}{default is 0.1. If the difference between r_sq from the k+1 breakpoint model
and the r_sq from the k breakpoint model is less than cutdiff, the optimal number
of breakpoint will be set as k instead of k+1.}

\item{num.try}{number of different seeds to try. If all num.try runs fails,
linear regression results will be returned (which represents one segment case).}

\item{keepfit}{whether to keep the fitted object.}

\item{forceRadj}{whether to use adjusted r_sq instead of r_sq in model comparisons.}

\item{NCores}{number of cores to use, default in detectCores() - 1.}

\item{makeShinyObject}{whether to output the object required for the SegReg shiny app.}

\item{outfile}{name and path of output of the object required for the shiny app.}
}
\value{
id.sign: direction of each sample based on pvalcut; -1: down, 0: no change, 1: up
slp: fitted slopes, slp.sign: sign of fitted slopes, slp.pval: p-value of each segment, 
bp: estimated breakpoints, fitted: fitted values radj: adjusted r_sq value of the model
fit: fit object
}
\description{
For each gene/feature, the function will fit maxk + 1 segmented regression models containing 0 - maxk breakpoints
(1 - (maxk+1)) segments. The model selection step selects the model with k breakpoints as that having both the highest r_sq value in addition to the difference in r_sq's from the k-1 and k breakpoint models being larger than cutdiff, i.e. r_sq(k) - r_sq(k-1) > cutdiff.
}
\examples{
d1 <- rbind(c(rep(1,50),1:50), c(100:1))
rownames(d1) <- c("g1","g2")
segreg(d1)
}
\author{
Rhonda Bacher & Ning Leng
}

