\name{SegRegExData}
\alias{SegRegExData}
\docType{data}
\title{
	Simulated gene level data set with 50 genes along 40 time points
}
\description{
	'SegRegExData' gives the gene level simulated data along 40 time points.
	It is used to demonstrate the SegReg package in breakpoint analysis of ordered expression data (e.g. time course, spatial course, etc.)
}
\format{
	SegRegExData is a matrix with 50 genes (rows) and 40 samples (columns). 
}
\seealso{
}
\examples{
data(SegRegExData)
str(SegRegExData)
}
\keyword{datasets}




