% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendheatmap.R
\name{trendheatmap}
\alias{trendheatmap}
\title{Draw heatmap of gene expression trends}
\usage{
trendheatmap(topres, showplot)
}
\arguments{
\item{topres}{results from topsegreg() function}

\item{showplot}{whether to show the heatmap.}

\item{savePDF}{save heatmap as a PDF.}

\item{outfile}{output file name; will be ignored if savePDF=F. Default is trend_heatmap.}

\item{pdfheight, pdfwidth}{height and width for the pdf file}

\item{sampnames}{names of samples for the columns of heatmap. Default is null, column names are not shown.}
}
\value{
The function takes significant genes called from 
the topsegreg() function. These genes are further grouped into three
groups : Up, Down or No Change (Same) in the first segment. Within each group,
the genes are sorted by their first break point. The heatmap shows expression trends of these 
three groups of genes. In the heatmap, red/blue/black represents up/down/nochange.
}
\author{
Ning Leng & Rhonda Bacher
}

