% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractpattern.R
\name{extractpattern}
\alias{extractpattern}
\title{Extract genes following a given pattern from segmented regression}
\usage{
extractpattern(segdata, radjcut = 10, pattern = NULL, delay=0)
}
\arguments{
\item{segdata}{segreg object}

\item{pattern}{vector containing pattern to search features. If length is one then only consider features with constant pattern across the entire time-course.}

\item{radjcut}{only consider features with adjusted r_sq > radjcut. Default = .5.}

\item{delay}{search for pattern starting after certain time-point.}
}
\value{
Gene: names of features containing pattern.

Breakpoint: estimated breakpoint connecting the segments of the pattern.
}
\examples{
extractpattern(segdata, pattern = c("up")) #increasing only features
extractpattern(segdata, pattern = c("up", "down")) #features with a peak
extractpattern(segdata, pattern = c("up", "down"), delay=20) #features with a peak after 20th time-point
extractpattern(segdata, pattern = c("same", "down")) #features which display delaying decreasing behavior.
}
\author{
Rhonda Bacher
}

