% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trendheatmap.R
\name{trendheatmap}
\alias{trendheatmap}
\title{Draw heatmap of gene expression trends}
\usage{
trendheatmap(topres, showplot)
}
\arguments{
\item{topres}{results from topsegreg() function}

\item{showplot}{whether show the heatmap}
}
\value{
The function takes significant genes called from
the topsegreg() function. These genes are further grouped into three
groups : go up/down or no change in the first segment. Within each group,
the genes are sorted by their first break point. The heatmap shows expression trends of these
three groups of genes. In the heatmap, red/blue/black represents up/down/nochange.
}
\description{
Draw heatmap of gene expression trends
}
\author{
Ning Leng
}

